+++ src/components/base/b-tree/b-tree.ts
@@ -76,6 +76,10 @@ class bTree extends bTreeProps implements iActiveItems, Foldable {
 
 	activeStore!: iActiveItems['activeStore'];
 
+	/** {@link Foldable.unfoldedStore} */
+	@system()
+	unfoldedStore: Foldable['unfoldedStore'] = new Set();
+
 	/** {@link iActiveItems.activeChangeEvent} */
 	@system()
 	readonly activeChangeEvent: string = 'change';
@@ -112,7 +116,7 @@ class bTree extends bTreeProps implements iActiveItems, Foldable {
 	}
 
 	/**
-	 * Stores bTree normalized items.
+	 * Stores `bTree` normalized items.
 	 * This store is needed because the `items` property should only be accessed via get/set.
 	 */
 	@field<bTree>((o) => o.sync.link<Item[]>((val) => {
@@ -333,7 +337,7 @@ class bTree extends bTreeProps implements iActiveItems, Foldable {
 	}
 
 	/**
-	 * True, if specified item has children
+	 * True, if the specified item has children
 	 * @param item
 	 */
 	protected hasChildren(item: this['Item']): boolean {
@@ -380,6 +384,10 @@ class bTree extends bTreeProps implements iActiveItems, Foldable {
 	 * @param item
 	 */
 	protected getFoldedPropValue(item: this['Item']): boolean {
+		if (this.unfoldedStore.has(item.value)) {
+			return false;
+		}
+
 		if (item.folded != null) {
 			return item.folded;
 		}
@@ -421,6 +429,10 @@ class bTree extends bTreeProps implements iActiveItems, Foldable {
 	/** {@link Values.initComponentValues} */
 	@hook('beforeDataCreate')
 	protected initComponentValues(itemsChanged: boolean = false): void {
+		if (itemsChanged) {
+			this.field.set('unfoldedStore', new Set());
+		}
+
 		this.values.init(itemsChanged);
 	}